We don't need uint64_t for len/cap in Move Slice.

+++ weak.h
@@ -119,7 +119,7 @@ struct CastleEventSlice {
 };
 
 struct MoveSlice {
-  uint64_t Len, Cap;
+  int Len, Cap;
   Move *Vals;
 };
